refactor(UI): ♻️ Standardized date handling & Updated timeline components.     -	Updated relative import paths in `Dates/Year.vue` to align with     the new directory structure.
    -	Refactored `Dates/Date.vue` by importing `DateItem` and
    `DateProps` from the type library and removing local interface
    definitions.

    -	Renamed the `Props` interface -> `DateProps` within
    `Dates/Date.vue` which improves naming consistency.

    -	Simplified temporal logic in `Dates/Date.vue` by replacing
    `dateObject` and `norwegianTime` with calls to the `setDateformat`
    function.

    -	Enhanced `Timeline/Card.vue` by adding a computed property for
    the `cls` prop and updating internal type references.

    -	Modified `preprosessor-utils.ts` in the `Composable` directory
    by updating type imports and allowing the return type to be
    `undefined` temporarily.

    -	Adjusted `MapTimeline` to include a blank caption, ensuring
    compatibility with updated caption handling in the figure components.
