+++ frontend/components/timeline/Card.vue
@@ -65,15 +65,11 @@
     
     //  --- Import & Props -setup logic
     import { computed } from 'vue';
-    import type { TimelineCardProps } from '~/types/props';
+    import type { TimelineCardProps } from '~/types/timeline';
 
-    const props = withDefaults(defineProps<TimelineCardProps>(),
-    {
-        cls: () => [],
-        isVisible: () => false,
-    });
+    const props = withDefaults(defineProps<TimelineCardProps>(), { cls: () => [], isVisible: () => false });
     
-    const cls = props.cls;
+    const cls = computed(() => props.cls);
     const data = computed(() => props.data);
     const emits = defineEmits(['toggleVisibility']);
