+++ frontend/types/date.d.ts
@@ -1,3 +1,4 @@
 export interface DateYearProps{ data: string; isVisible: boolean; }
-export interface DateObject { created: string; end?: string | null; updated?: string | null; }
-export interface DateItem { time?: string | null; date?: string | null; text?: string; delimiter?: string; updated?: Date | string | null; }
+interface DateProps { cls?: Array<string>; data:DateItem; isArticle?: boolean; };
+export interface DateObject { created: string;  end?: string | null; date: Record }
+export interface DateItem { current?: string ; updated?: string | null; end?: string; time?: string | null; date?: string | null ; text?: string; delimiter?: string; type?: string;}
